requirements.txt
streamlit
requests
feedparser
pandas
openai
google-generativeai
